kernel32-Processor_Group: Remove patch.
This seems to be another one of those cases where Sebastian added a patch for an
application [1], then never submitted the patch upstream, and then someone else
submitted a different patch upstream [2], but Sebastian didn't bother checking
if his version was actually still necessary, and instead assumed it was (or just
decided it was better) and rebased the patch [3], and of course still never
submitted it upstream.

In this case the patch is for Terragen 4. That application does not, according
to my testing, actually depend on success from SetThreadIdealProcessorEx().

The patch does not add a correct implementation, is trivial to recreate, and is
easy to re-debug (the function still prints a FIXME), so it's not adding any
value. Remove it.

[1] https://gitlab.winehq.org/wine/wine-staging/-/commit/a12dca03ce2c24baf73d296e0d391fd0a67842db

[2] https://source.winehq.org/git/wine.git/commitdiff/980754bff773402efb52457b4536368900f3a1a3

[3] https://gitlab.winehq.org/wine/wine-staging/-/commit/0c46d1e8a20f3d85ad9e3f91b7b71a67fc10377b