Updated odbc-remove-unixodbc patchset
Functions SQLDataSources*/SQLDrivers* are local to this DLL and not forwards onto the driver DLL.
Hense the Environment handle being passed as the first parameter.

Said functions need to return SQL_NO_DATA to state we have no data instead of an SQL_ERROR.

Fixes: https://bugs.winehq.org/show_bug.cgi?id=56616

Allows it to run now, more work will be required to make it fully functional.