Updated cmd-launch-association patchset
Fix the compile error or "Invalid pointer type passed to function"

Function GetExitCodeProcess, takes a DWORD as a section paramater,
and the code for the used errorlevel (aka an int).

The old code, used a local for the return value and reassigned it.
I've taken the approach to just change the errorlevel type.